refactor(frontend): extract REASON_HINTS to its own module
