+++ frontend/src/components/PreValidationResultsModal.tsx
@@ -17,6 +17,7 @@ import type {
   EntirelyImpossibleMpCamper,
 } from '../services/solver'
 import { CamperNameButton } from './impossibility/CamperNameButton'
+import { camperActionHints } from './impossibility/reasonHints'
 
 const CamperDetailsPanel = lazy(() => import('./CamperDetailsPanel'))
 
@@ -149,25 +150,6 @@ const FRIENDLY_REASON_LABELS: Record<string, string> = {
 
 function friendlyReasonLabel(code: string): string {
   return FRIENDLY_REASON_LABELS[code] || code
-}
-
-// Per-reason hint copy: tells staff what to actually do for each rejection.
-// Exported so PostValidationResultsModal can reuse the same map.
-export const REASON_HINTS: Record<string, string> = {
-  target_not_in_solver: 'check enrollment — requested camper not on roster',
-  grade_compatibility: 'grade gap too wide — confirm priority with the family',
-  cross_session: 'requested friend is in a different session — confirm intent',
-  pair_no_shared_bunk: 'cross-gender request — confirm with the family',
-  age_pref_no_eligible_grade: 'at the youngest/oldest grade — preference is moot',
-  malformed: 'request is missing a name — needs parent resubmission',
-}
-
-function camperActionHints(reasonCodes: string[]): string {
-  const hints = new Set<string>()
-  for (const code of reasonCodes) {
-    hints.add(REASON_HINTS[code] ?? 'review request')
-  }
-  return Array.from(hints).join(' / ')
 }
 
 function EntirelyImpossibleMpSection({
